pyproject.toml
[tool.poetry]
name = "resize-photo"
version = "0.1.0"
description = ""
authors = ["sontyubo <sontyubo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.4.0"
natsort = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
